--- src/main/java/Engulfy/Error/EngulfyError.java
@@ -1,7 +1,0 @@
-package Engulfy.Error;
-
-public class EngulfyError extends Exception{
-    public EngulfyError(String message){
-        super(message);
-    }
-}
