--- src/main/java/Engulfy/Task/Task.java
@@ -1,28 +1,0 @@
-package Engulfy.Task;
-
-public class Task {
-    protected String description;
-    public boolean isDone;
-
-    public Task(String description) {
-        this.description = description;
-        this.isDone = false;
-    }
-
-    public String getStatusIcon() {
-        return (isDone ? "[X]" : "[ ]");
-    }
-
-    public void markAsDone() {
-        isDone = true;
-    }
-
-    public void markAsNotDone() {
-        isDone = false;
-    }
-
-    @Override
-    public String toString() {
-        return getStatusIcon() + " " + description;
-    }
-}
