--- src/main/java/Engulfy/Task/Todo.java
@@ -1,12 +1,0 @@
-package Engulfy.Task;
-
-public class Todo extends Task {
-    public Todo(String description) {
-        super(description);
-    }
-
-    @Override
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-}
