+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 ____________________________________________________________
-Hello! I'm Engulfy
+Hello! I'm engulfy
 What can I do for you?
 ____________________________________________________________
 Got it. I've added this task:
